Crate tm4c_hal

Generic implementation code for both TM4C123 and TM4C129.

Modules

bb

Code to handle bit-banding.

delay

Code for busy-waiting

gpio

Code for GPIO pins

i2c

Common I2C code for TM4C123 and TM4C129

serial

Serial code that is generic to both the TM4C123 and TM4C129, such as the pin traits.

sysctl

Code for the System Control module.

time

Time units

Macros

gpio_macro

! An internal macro to implement the GPIO functionality for each port

i2c_busy_wait

Spins until the controler is ready (mcs.busy is clear) and optionally on another field of the mcs register until it is clear or set (depending on op parameter).

i2c_hal

Implements embedded-hal for an TM4C I2C peripheral

i2c_pins

Implements the traits for an I2C peripheral

uart_hal_macro

! An internal macro to implement the UART functionality for each peripheral

uart_pin_macro

! An internal macro to help define all the different pin typestates